+++ click_extended/decorators/convert/__init__.py
@@ -3,6 +3,7 @@
 from click_extended.decorators.convert.convert_area import convert_area
 from click_extended.decorators.convert.convert_bits import convert_bits
 from click_extended.decorators.convert.convert_distance import convert_distance
+from click_extended.decorators.convert.convert_energy import convert_energy
 from click_extended.decorators.convert.convert_pressure import convert_pressure
 from click_extended.decorators.convert.convert_speed import convert_speed
 from click_extended.decorators.convert.convert_temperature import (
@@ -16,6 +17,7 @@ __all__ = [
     "convert_area",
     "convert_bits",
     "convert_distance",
+    "convert_energy",
     "convert_pressure",
     "convert_time",
     "convert_speed",